pyproject.toml
[project]
name = "nexus-study-backend"
version = "0.1.0"
description = "FastAPI backend for nexus-study"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "python-dotenv>=1.0.1",
  "redis>=5.0.8",
  "supabase>=2.0.0",
  "httpx>=0.27.0",
]

[dependency-groups]
dev = [
  "pytest>=8.0.0",
  "httpx>=0.27.0",
  "ruff>=0.5.0",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
